backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1
greenlet==3.1.1

# LLM Providers
openai==1.82.0
cerebras_cloud_sdk>=1.67.0
httpx==0.28.1

# Docker
docker==7.1.0

# Git
gitpython==3.1.44

# Aider
aider-chat==0.72.1

# WebSocket
websockets==14.1

# Utilities
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
